+++ frontend/src/components/Login/Login.tsx
@@ -1,15 +1,40 @@
+import { useReactiveVar } from '@apollo/client';
 import { Button } from '@mui/material';
 import { Box } from '@mui/system';
+import { useState } from 'react';
+import { Redirect } from 'react-router';
 
-interface LoginProps {
-  onClickLoginButton: (oauthStrategy: 'google' | '42') => void;
-  buttonsEnabledState: boolean;
-}
+import { userIdVar } from '../..';
 
-function Login({
-  onClickLoginButton,
-  buttonsEnabledState,
-}: LoginProps): JSX.Element {
+function Login(): JSX.Element {
+  const userId = useReactiveVar(userIdVar);
+
+  const [buttonsEnabledState, setButtonsEnabledState] = useState<boolean>(true);
+
+  const onClickLoginButton = (oauthStrategy: 'google' | '42'): void => {
+    if (
+      process.env.REACT_APP_BACKEND_HOST &&
+      process.env.REACT_APP_BACKEND_PORT
+    ) {
+      setButtonsEnabledState(false);
+
+      const loginURI = `http://${process.env.REACT_APP_BACKEND_HOST}:${process.env.REACT_APP_BACKEND_PORT}/auth/login/${oauthStrategy}`;
+      // popupWindow = popupWindowCenter(
+      //   loginURI,
+      //   `Login with ${oauthStrategy}`,
+      //   400,
+      //   800
+      // );
+      location.replace(loginURI);
+    } else {
+      throw `Undefined environment variables: ${
+        process.env.REACT_APP_BACKEND_HOST ? '' : 'REACT_APP_BACKEND_HOST'
+      } ${process.env.REACT_APP_BACKEND_PORT ? '' : 'REACT_APP_BACKEND_PORT'}`;
+    }
+  };
+
+  if (userId) return <Redirect to="/" />;
+
   return (
     <Box width="250px" sx={{ display: 'flex', flexDirection: 'column' }}>
       <Button
