+++ frontend/src/components/Login/index.ts
@@ -1,3 +1,3 @@
-import LoginContainer from './LoginContainer';
+import Login from './Login';
 
-export default LoginContainer;
+export default Login;
